crew_backend/requirements.txt
crewai==0.80.0
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
pydantic==2.7.0
